mcp/todoist/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
markdown-it-py==3.0.0
mcp==1.6.0
mdurl==0.1.2
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.8.1
pygments==2.19.1
python-dotenv==1.1.0
requests==2.32.3
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.2
todoist-api-python==2.1.7
typer==0.15.2
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.1
